+++ e2e/fixtures/citizen/createCase/C100/people/provideDetailsContent.ts
