+++ e2e/journeys/citizen/createCase/C100People/C100People1.ts
@@ -1,8 +1,8 @@
 import { Page } from "@playwright/test";
 import { UserRole } from "../../../../common/types";
-import { AddChildrenPage } from "../../../../pages/citizen/createCase/C100/people1/addChildrenPage";
-import { ProvideDetailsPage } from "../../../../pages/citizen/createCase/C100/people1/provideDetailsPage";
-import { WhichDecisionsPage } from "../../../../pages/citizen/createCase/C100/people1/whichDecisionsPage";
+import { AddChildrenPage } from "../../../../pages/citizen/createCase/C100/people/addChildrenPage";
+import { ProvideDetailsPage } from "../../../../pages/citizen/createCase/C100/people/provideDetailsPage";
+import { WhichDecisionsPage } from "../../../../pages/citizen/createCase/C100/people/whichDecisionsPage";
 import { ApplicantGender } from "../../../../common/types";
 
 interface C100People1Options {
