+++ e2e/pages/citizen/createCase/C100/people/addChildrenPage.ts
@@ -1,11 +1,11 @@
 import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper";
 import { Page } from "@playwright/test";
 import { Selectors } from "../../../../../common/selectors";
-import { WhichDecisionsContent } from "../../../../../fixtures/citizen/createCase/C100/people1/whichDecisionsContent";
+import { AddChildrenContent } from "../../../../../fixtures/citizen/createCase/C100/people/addChildrenContent";
 import { Helpers } from "../../../../../common/helpers";
 import { CommonStaticText } from "../../../../../common/commonStaticText";
 
-interface WhichDecisionsPageOptions {
+interface AddChildrenPageOptions {
   page: Page;
   accessibilityTest: boolean;
   errorMessaging: boolean;
@@ -20,30 +20,17 @@ interface fillInFieldsOptions {
   page: Page;
 }
 
-enum checkBoxIds {
-  needsResolution1 = "needsResolution",
-  needsResolution2 = "needsResolution-2",
-  needsResolution3 = "needsResolution-3",
-  needsResolution4 = "needsResolution-4",
-  needsResolution5 = "needsResolution-5",
-  needsResolution6 = "needsResolution-6",
-  needsResolution7 = "needsResolution-7",
-  needsResolution8 = "needsResolution-8",
-  needsResolution9 = "needsResolution-9",
-  needsResolution10 = "needsResolution-10",
-  needsResolution11 = "needsResolution-11",
-  needsResolution12 = "needsResolution-12",
-  needsResolution13 = "needsResolution-13",
-  needsResolution14 = "needsResolution-14",
-  needsResolution15 = "needsResolution-15",
+enum inputIds {
+  firstName = "#c100TempFirstName",
+  lastName = "#c100TempLastName",
 }
 
-export class WhichDecisionsPage {
-  public static async whichDecisionsPage({
+export class AddChildrenPage {
+  public static async addChildrenPage({
     page: page,
     accessibilityTest: accessibilityTest,
     errorMessaging: errorMessaging,
-  }: WhichDecisionsPageOptions): Promise<void> {
+  }: AddChildrenPageOptions): Promise<void> {
     await this.checkPageLoads({
       page: page,
       accessibilityTest: accessibilityTest,
@@ -61,20 +48,30 @@ export class WhichDecisionsPage {
     accessibilityTest: accessibilityTest,
   }: checkPageLoadsOptions): Promise<void> {
     await page.waitForSelector(
-      `${Selectors.GovukHeadingXL}:text-is("${WhichDecisionsContent.pageTitle}")`,
+      `${Selectors.GovukHeadingXL}:text-is("${AddChildrenContent.pageTitle}")`,
     );
     await Promise.all([
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukHint}:text-is("${WhichDecisionsContent.hint}")`,
+        `${Selectors.GovukInsetText}:text-is("${AddChildrenContent.insetText}")`,
+        1,
+      ),
+      Helpers.checkVisibleAndPresent(
+        page,
+        `${Selectors.h1}:text-is("${AddChildrenContent.heading}")`,
         1,
       ),
       Helpers.checkGroup(
         page,
-        15,
-        WhichDecisionsContent,
+        2,
+        AddChildrenContent,
         "label",
         `${Selectors.GovukLabel}`,
+      ),
+      Helpers.checkVisibleAndPresent(
+        page,
+        `${Selectors.GovukHint}:text-is("${AddChildrenContent.hint}")`,
+        1,
       ),
     ]);
     if (accessibilityTest) {
@@ -86,19 +83,25 @@ export class WhichDecisionsPage {
     await page.click(
       `${Selectors.button}:text-is("${CommonStaticText.paddedContinue}")`,
     );
-    await page.waitForSelector(
-      `${Selectors.GovukErrorSummaryTitle}:text-is("${CommonStaticText.errorSummaryTitle}")`,
-    );
     await Promise.all([
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.a}:text-is("${WhichDecisionsContent.errorLink}")`,
+        `${Selectors.GovukErrorSummaryTitle}:text-is("${CommonStaticText.errorSummaryTitle}")`,
         1,
       ),
-      Helpers.checkVisibleAndPresent(
+      Helpers.checkGroup(
         page,
-        `${Selectors.ErrorMessage}:text-is("${WhichDecisionsContent.errorLink}")`,
-        1,
+        2,
+        AddChildrenContent,
+        "errorLink",
+        `${Selectors.a}`,
+      ),
+      Helpers.checkGroup(
+        page,
+        2,
+        AddChildrenContent,
+        "errorLink",
+        `${Selectors.ErrorMessage}`,
       ),
     ]);
   }
@@ -106,8 +109,8 @@ export class WhichDecisionsPage {
   private static async fillInFields({
     page: page,
   }: fillInFieldsOptions): Promise<void> {
-    for (const selector of Object.values(checkBoxIds)) {
-      await page.click(selector);
+    for (const selector of Object.values(inputIds)) {
+      await page.fill(`${selector}`, AddChildrenContent.exampleText);
     }
     await page.click(
       `${Selectors.button}:text-is("${CommonStaticText.paddedContinue}")`,
