+++ e2e/pages/citizen/createCase/C100/people/whichDecisionsPage.ts
@@ -1,11 +1,11 @@
 import AccessibilityTestHelper from "../../../../../common/accessibilityTestHelper";
 import { Page } from "@playwright/test";
 import { Selectors } from "../../../../../common/selectors";
-import { AddChildrenContent } from "../../../../../fixtures/citizen/createCase/C100/people1/addChildrenContent";
+import { WhichDecisionsContent } from "../../../../../fixtures/citizen/createCase/C100/people/whichDecisionsContent";
 import { Helpers } from "../../../../../common/helpers";
 import { CommonStaticText } from "../../../../../common/commonStaticText";
 
-interface AddChildrenPageOptions {
+interface WhichDecisionsPageOptions {
   page: Page;
   accessibilityTest: boolean;
   errorMessaging: boolean;
@@ -20,17 +20,30 @@ interface fillInFieldsOptions {
   page: Page;
 }
 
-enum inputIds {
-  firstName = "#c100TempFirstName",
-  lastName = "#c100TempLastName",
+enum checkBoxIds {
+  needsResolution1 = "needsResolution",
+  needsResolution2 = "needsResolution-2",
+  needsResolution3 = "needsResolution-3",
+  needsResolution4 = "needsResolution-4",
+  needsResolution5 = "needsResolution-5",
+  needsResolution6 = "needsResolution-6",
+  needsResolution7 = "needsResolution-7",
+  needsResolution8 = "needsResolution-8",
+  needsResolution9 = "needsResolution-9",
+  needsResolution10 = "needsResolution-10",
+  needsResolution11 = "needsResolution-11",
+  needsResolution12 = "needsResolution-12",
+  needsResolution13 = "needsResolution-13",
+  needsResolution14 = "needsResolution-14",
+  needsResolution15 = "needsResolution-15",
 }
 
-export class AddChildrenPage {
-  public static async addChildrenPage({
+export class WhichDecisionsPage {
+  public static async whichDecisionsPage({
     page: page,
     accessibilityTest: accessibilityTest,
     errorMessaging: errorMessaging,
-  }: AddChildrenPageOptions): Promise<void> {
+  }: WhichDecisionsPageOptions): Promise<void> {
     await this.checkPageLoads({
       page: page,
       accessibilityTest: accessibilityTest,
@@ -48,30 +61,20 @@ export class AddChildrenPage {
     accessibilityTest: accessibilityTest,
   }: checkPageLoadsOptions): Promise<void> {
     await page.waitForSelector(
-      `${Selectors.GovukHeadingXL}:text-is("${AddChildrenContent.pageTitle}")`,
+      `${Selectors.GovukHeadingXL}:text-is("${WhichDecisionsContent.pageTitle}")`,
     );
     await Promise.all([
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukInsetText}:text-is("${AddChildrenContent.insetText}")`,
-        1,
-      ),
-      Helpers.checkVisibleAndPresent(
-        page,
-        `${Selectors.h1}:text-is("${AddChildrenContent.heading}")`,
+        `${Selectors.GovukHint}:text-is("${WhichDecisionsContent.hint}")`,
         1,
       ),
       Helpers.checkGroup(
         page,
-        2,
-        AddChildrenContent,
+        15,
+        WhichDecisionsContent,
         "label",
         `${Selectors.GovukLabel}`,
-      ),
-      Helpers.checkVisibleAndPresent(
-        page,
-        `${Selectors.GovukHint}:text-is("${AddChildrenContent.hint}")`,
-        1,
       ),
     ]);
     if (accessibilityTest) {
@@ -83,25 +86,19 @@ export class AddChildrenPage {
     await page.click(
       `${Selectors.button}:text-is("${CommonStaticText.paddedContinue}")`,
     );
+    await page.waitForSelector(
+      `${Selectors.GovukErrorSummaryTitle}:text-is("${CommonStaticText.errorSummaryTitle}")`,
+    );
     await Promise.all([
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukErrorSummaryTitle}:text-is("${CommonStaticText.errorSummaryTitle}")`,
+        `${Selectors.a}:text-is("${WhichDecisionsContent.errorLink}")`,
         1,
       ),
-      Helpers.checkGroup(
+      Helpers.checkVisibleAndPresent(
         page,
-        2,
-        AddChildrenContent,
-        "errorLink",
-        `${Selectors.a}`,
-      ),
-      Helpers.checkGroup(
-        page,
-        2,
-        AddChildrenContent,
-        "errorLink",
-        `${Selectors.ErrorMessage}`,
+        `${Selectors.ErrorMessage}:text-is("${WhichDecisionsContent.errorLink}")`,
+        1,
       ),
     ]);
   }
@@ -109,8 +106,8 @@ export class AddChildrenPage {
   private static async fillInFields({
     page: page,
   }: fillInFieldsOptions): Promise<void> {
-    for (const selector of Object.values(inputIds)) {
-      await page.fill(`${selector}`, AddChildrenContent.exampleText);
+    for (const selector of Object.values(checkBoxIds)) {
+      await page.click(selector);
     }
     await page.click(
       `${Selectors.button}:text-is("${CommonStaticText.paddedContinue}")`,
